Sql Patch For Userinfo {Axel}
Lets see What happen Fuck python 3.9.9


SIg by 👎  aksraashish@vex.me

+++ vexana/modules/sql/userinfo_sql.py
@@ -1,11 +1,12 @@
 import threading
+
 from vexana.modules.sql import BASE, SESSION
-from sqlalchemy import Column, UnicodeText, BigInteger
+from sqlalchemy import Column, Integer, UnicodeText
 
 
 class UserInfo(BASE):
     __tablename__ = "userinfo"
-    user_id = Column(BigInteger, primary_key=True)
+    user_id = Column(Integer, primary_key=True)
     info = Column(UnicodeText)
 
     def __init__(self, user_id, info):
@@ -18,7 +19,7 @@ class UserInfo(BASE):
 
 class UserBio(BASE):
     __tablename__ = "userbio"
-    user_id = Column(BigInteger, primary_key=True)
+    user_id = Column(Integer, primary_key=True)
     bio = Column(UnicodeText)
 
     def __init__(self, user_id, bio):
